[QHC-662] Implement the Experiment class (#782)
* initial WIP

* implement StructuredProgram and Experiment classes

* fix Parameter enum

* Refactor and fix code quality

* add changelog entries

* fix unused import
