+++ tests/qprogram/test_qprogram.py
@@ -28,6 +28,9 @@ from qililab.qprogram.operations import (
 )
 from qililab.qprogram.variable import FloatVariable, IntVariable
 from qililab.utils.serialization import deserialize, deserialize_from, serialize, serialize_to
+from tests.qprogram.test_structured_program import (  # pylint: disable=no-name-in-module, import-error
+    TestStructuredProgram,
+)
 
 
 @pytest.fixture(name="sample_qprogram_string")
@@ -66,38 +69,12 @@ def get_sample_qprogram_string():
 
 
 # pylint: disable=maybe-no-member, protected-access
-class TestQProgram:
+class TestQProgram(TestStructuredProgram):
     """Unit tests checking the QProgram attributes and methods"""
 
-    def test_init(self):
-        """Test init method"""
-        qp = QProgram()
-        assert isinstance(qp._body, Block)
-        assert len(qp._body.elements) == 0
-        assert isinstance(qp._block_stack, deque)
-        assert len(qp._block_stack) == 1
-        assert isinstance(qp._variables, list)
-        assert len(qp._variables) == 0
-
-    def test_str_method(self, sample_qprogram_string):
-        assert str(sample_qprogram_string[0]) == sample_qprogram_string[1]
-
-    def test_active_block_property(self):
-        """Test _active_block property"""
-        qp = QProgram()
-        assert isinstance(qp._active_block, Block)
-        assert qp._active_block is qp._body
-
-    def test_block_method(self):
-        """Test block method"""
-        qp = QProgram()
-        with qp.block() as block:
-            # __enter__
-            assert isinstance(block, Block)
-            assert qp._active_block is block
-        # __exit__
-        assert len(qp._body.elements) == 1
-        assert qp._body.elements[0] is block
+    @pytest.fixture
+    def instance(self):
+        return QProgram()
 
     def test_with_bus_mapping_method(self):
         """Test with_bus_mapping method"""
@@ -227,74 +204,7 @@ class TestQProgram:
         assert isinstance(new_qp.body.elements[0].elements[7], Measure)
         assert isinstance(new_qp.body.elements[0].elements[8], Measure)
 
-    def test_infinite_loop_method(self):
-        """Test infinite_loop method"""
-
-        qp = QProgram()
-        with qp.infinite_loop() as loop:
-            # __enter__
-            assert isinstance(loop, InfiniteLoop)
-            assert qp._active_block is loop
-        # __exit__
-        assert len(qp._body.elements) == 1
-        assert qp._body.elements[0] is loop
-        assert qp._active_block is qp._body
-
-    def test_parallel_method(self):
-        """Test parallel method"""
-        qp = QProgram()
-        var1 = qp.variable(Domain.Scalar, int)
-        var2 = qp.variable(Domain.Scalar, float)
-        with qp.parallel(
-            loops=[
-                ForLoop(variable=var1, start=0, stop=10, step=1),
-                ForLoop(variable=var2, start=0.0, stop=1.0, step=0.1),
-            ]
-        ) as loop:
-            # __enter__
-            assert isinstance(loop, Parallel)
-            assert len(loop.loops) == 2
-            assert qp._active_block is loop
-        # __exit__
-        assert len(qp._body.elements) == 1
-        assert qp._body.elements[0] is loop
-        assert qp._active_block is qp._body
-
-    def test_for_loop_method(self):
-        """Test loop method"""
-        qp = QProgram()
-        variable = qp.variable(Domain.Scalar, int)
-        start, stop, step = 0, 100, 5
-        with qp.for_loop(variable=variable, start=start, stop=stop, step=step) as loop:
-            # __enter__
-            assert isinstance(loop, ForLoop)
-            assert loop.variable == variable
-            assert loop.start == start
-            assert loop.stop == stop
-            assert loop.step == step
-            assert qp._active_block is loop
-        # __exit__
-        assert len(qp._body.elements) == 1
-        assert qp._body.elements[0] is loop
-        assert qp._active_block is qp._body
-
-    def test_loop_method(self):
-        """Test loop method"""
-        qp = QProgram()
-        variable = qp.variable(Domain.Scalar, int)
-        values = np.ones(10, dtype=int)
-        with qp.loop(variable=variable, values=values) as loop:
-            # __enter__
-            assert isinstance(loop, Loop)
-            assert loop.variable == variable
-            assert np.array_equal(loop.values, values)
-            assert qp._active_block is loop
-        # __exit__
-        assert len(qp._body.elements) == 1
-        assert qp._body.elements[0] is loop
-        assert qp._active_block is qp._body
-
-    def test_acquire_loop_method(self):
+    def test_average_method(self):
         """Test acquire_loop method"""
         qp = QProgram()
         with qp.average(shots=1000) as loop:
@@ -446,64 +356,6 @@ class TestQProgram:
 
         with pytest.raises(AttributeError):
             qp.qblox.set_markers(bus="drive", mask="0111011")
-
-    def test_variable_method(self):
-        """Test variable method"""
-        qp = QProgram()
-        frequency_variable = qp.variable(Domain.Frequency)
-        phase_variable = qp.variable(Domain.Phase)
-        voltage_variable = qp.variable(Domain.Voltage)
-        time_variable = qp.variable(Domain.Time)
-        int_scalar_variable = qp.variable(Domain.Scalar, int)
-        float_scalar_variable = qp.variable(Domain.Scalar, float)
-
-        # Test instantiation
-        assert isinstance(frequency_variable, float)
-        assert isinstance(frequency_variable, FloatVariable)
-        assert frequency_variable.domain is Domain.Frequency
-        assert frequency_variable.value is None
-
-        assert isinstance(phase_variable, float)
-        assert isinstance(phase_variable, FloatVariable)
-        assert phase_variable.domain is Domain.Phase
-        assert phase_variable.value is None
-
-        assert isinstance(voltage_variable, float)
-        assert isinstance(voltage_variable, FloatVariable)
-        assert voltage_variable.domain is Domain.Voltage
-        assert voltage_variable.value is None
-
-        assert isinstance(time_variable, int)
-        assert isinstance(time_variable, IntVariable)
-        assert time_variable.domain is Domain.Time
-        assert time_variable.value is None
-
-        assert isinstance(int_scalar_variable, int)
-        assert isinstance(int_scalar_variable, IntVariable)
-        assert int_scalar_variable.domain is Domain.Scalar
-        assert int_scalar_variable.value is None
-
-        assert isinstance(float_scalar_variable, float)
-        assert isinstance(float_scalar_variable, FloatVariable)
-        assert float_scalar_variable.domain is Domain.Scalar
-        assert float_scalar_variable.value is None
-
-        # Test storing in QProgram's _variables
-        assert len(qp._variables) == 6
-
-    def test_variable_method_raises_error_if_domain_is_scalar_and_type_is_none(self):
-        """Test variable method"""
-        qp = QProgram()
-        with pytest.raises(ValueError, match="You must specify a type in a scalar variable."):
-            qp.variable(Domain.Scalar)
-
-    def test_variable_method_raises_error_if_domain_is_not_scalar_and_type_is_set(self):
-        """Test variable method"""
-        qp = QProgram()
-        with pytest.raises(
-            ValueError, match="When declaring a variable of a specific domain, its type is inferred by its domain."
-        ):
-            qp.variable(Domain.Frequency, int)
 
     def test_operation_with_variable_of_wrong_domain_raises_error(self):
         """Test that any operation when used with a variable of wrong domain raises an error."""
